[make] fixed warnings

+++ src/reader.c
@@ -222,7 +222,7 @@ struct s_mat * dataRead(char * start) {
 							// test strtol
 							errno = 0;
 							matStruct->matTab[i + j][ligne][colonne] = strtol(end, &end, 10);
-							if ((matStruct->matTab[i + j][ligne][colonne] == LONG_MAX || matStruct->matTab[ligne][colonne] == LONG_MIN) && errno == ERANGE) {
+							if ((matStruct->matTab[i + j][ligne][colonne] == LONG_MAX || matStruct->matTab[ligne][colonne] == (int) LONG_MIN) && errno == ERANGE) {
 								perror("strtol");
 								return NULL;
 							}
@@ -240,7 +240,7 @@ struct s_mat * dataRead(char * start) {
 						// test strtol
 						errno = 0;
 						matStruct->matTab[i + j][ligne][colonne] = strtol(end, &end, 10);
-						if ((matStruct->matTab[i + j][ligne][colonne] == LONG_MAX || matStruct->matTab[ligne][colonne] == LONG_MIN) && errno == ERANGE) {
+						if ((matStruct->matTab[i + j][ligne][colonne] == LONG_MAX || matStruct->matTab[ligne][colonne] == (int) LONG_MIN) && errno == ERANGE) {
 							perror("strtol");
 							return NULL;
 						}
